move ptr to the intended place
